feat: :sparkles: Creation of unitOfWork
The Unit of Work pattern is used to manage transactions and the changes made to a set of objects (usually related to a database) during the execution of a software application. It ensures that all changes are committed as a single transaction, either all succeeding or all failing, thus maintaining the integrity of the data.